Boarding 'origin/bb' on Release Train - Merging into release-train Branch (#9)
* fix(on-pr-merge): add secrets.GITHUB_TOKEN (offered out-of-the-box by GH) to 'gh pr' env

* fix(on-pr-merge): pass a PAT as GITHUB_TOKEN

Pass a PAT, with 'Read and Write access to actions, code, and pull requests' and 'Read access to metadata', as value to the GITHUB_TOKEN env var declared solely in the step that runs 'gh pr create' and is opens a Github Pull Request.

* fix(on-pr-merge): pass a PAT as GH_TOKEN

Pass a PAT, with 'Read and Write access to actions, code, and pull requests' and 'Read access to metadata', as value to the GH_TOKEN env var declared solely in the step that runs 'gh pr merge "${{ env.BOARDING }}" --auto --delete-branch --squash ' and Enables Auto Merge on the Pull Request, opened in previous step.

* dummy commit for testing purposes